19 june

+++ src/main/java/repository/WriterRepository.java
@@ -1,4 +1,7 @@
 package repository;
 
-public interface WriterRepository extends GenericRepository<Writer, Long>{
+import model.Writer;
+import repository.Impl.GenericRepository;
+
+public interface WriterRepository extends GenericRepository<Writer, Integer> {
 }
